Start work on creating a dashboard to display sensor values.
The web server now provides XML data from the sensors that can be queried by
JavaScript.  The presentation can now be controlled by the HTML, CSS, and
JavaScript files.  In the process, a bug in the svg thermometer code was
discovered and corrected.  This needs to also be pushed to the generic web
server repository.