Change handler task system
The original system with the handler tasks being detached and left to run to
completion after being started did not work as a long term solution.  It
appears that the tasks were not being cleaned up by the run time system.  This
led to a memory leak.  The new system creates a limited number of tasks to
start with and then allocates work to them on a round-robin basis.  This seems
to have cured the memory leak.