Make RS-485 data protected and prepare to send commands
The RS-485 data store have been converted to a protected type to help avoid
any problems when trying to read the data at the same time as the state
machine task is writing data.

Add code to transmit text on the RS-485 port.  This will be used to send
commands to the controller Arduino.  Note that send data goes on an
entirely different internal RS-485 bus and is not read by any of the other
devices on the RS-485 network.