Work on exception handling and robustness.
Starting the two RS-485 processing tasks seems to be an area of some
sensitivity.  It seems that if things aren't done just right, one of
the tasks will fail to open its port properly without an exception on
the open statement.  This needs some more research to correctly
characterize and solve the problem.