Change command task name and other changes
Change rs485_cmd_type to command_task, which better describes what it really
is.
Change Char_IO to Ada.Text_IO for character input.  This is not a completely
transparent change since LF characters are now swallowed.  As a result, the
state machine looks for a CR to end the line.
Do a bit more checking to try and identify cause of problems opening the
RS-485 port as input and as output.  No success yet, but making sure that the
two opens happen one after another doesn't seem to fix the problem.